+++ packages/antwerp-ui/react-components/src/utils/time.utils.ts
@@ -1,4 +1,4 @@
-import { isAfter, isBefore, isSameDay } from 'date-fns';
+import { isAfter, isBefore, isSameDay, format, Locale } from 'date-fns';
 
 export function isInRange(date: Date, from?: string, to?: string, list?: string[]): boolean {
   return (
@@ -6,4 +6,20 @@ export function isInRange(date: Date, from?: string, to?: string, list?: string[
     (from && (isAfter(date, new Date(from)) || isSameDay(date, new Date(from)))) ||
     !!list?.find((u) => isSameDay(date, new Date(u)))
   );
+}
+
+export function formatIfValid(date: string, dateFormat: string) {
+  try {
+    return format(new Date(date), dateFormat);
+  } catch (_e) {
+    return date;
+  }
+}
+
+export function formatWithFallback(date: Date, dateFormat: string, locale?: Locale, fallback = '') {
+  try {
+    return format(date, dateFormat, { locale: locale });
+  } catch (_e) {
+    return fallback;
+  }
 }