+++ CHANGELOG.md
@@ -7,6 +7,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Datepicker doesn't crash when providing it an invalid date as value
+- Controlled Calendar doesn't crash when providing it an invalid date as value
+
+### Changed
+
+- Datepicker `onChange` function is now triggered for invalid dates (return empty string)
+- Datepicker `onChange` function returns the input string value as second parameter
+
+### Added
+
+- Datepicker `errorMsgFunction` to fully customize the error based on the input
+
 ## 7.1.1 - 2024-03-07
 
 ### Added